+++ ace/Select_Reactor_Base.cpp
@@ -1,4 +1,4 @@
-// $Id: Select_Reactor_Base.cpp 96985 2013-04-11 15:50:32Z huangh $
+// $Id: Select_Reactor_Base.cpp 97246 2013-08-07 07:10:20Z johnnyw $
 
 #include "ace/Select_Reactor_Base.h"
 #include "ace/Reactor.h"
@@ -892,7 +892,7 @@ ACE_Select_Reactor_Notify::read_notify_pipe (ACE_HANDLE handle,
   if (n > 0)
     {
       // Check to see if we've got a short read.
-      if (n != sizeof buffer)
+      if ((size_t)n != sizeof buffer)
         {
           ssize_t const remainder = sizeof buffer - n;
